[WebDriver][win] separate AutomationSessionClient and AutomationClient into different files https://bugs.webkit.org/show_bug.cgi?id=290982
Reviewed by Fujii Hironori.

Currently, `AutomationSessionClient` and `AutomationClient` are
implemented in the same file on the win port.
They should be separated into different files.

* Source/WebKit/PlatformWin.cmake:
* Source/WebKit/UIProcess/win/AutomationClientWin.cpp:
(WebKit::AutomationSessionClient::AutomationSessionClient): Deleted.
(WebKit::AutomationSessionClient::close): Deleted.
(WebKit::AutomationSessionClient::didReceiveAuthenticationChallenge): Deleted.
(WebKit::AutomationSessionClient::requestNewPageWithOptions): Deleted.
(WebKit::AutomationSessionClient::didDisconnectFromRemote): Deleted.
(WebKit::AutomationSessionClient::retainWebView): Deleted.
(WebKit::AutomationSessionClient::releaseWebView): Deleted.
* Source/WebKit/UIProcess/win/AutomationClientWin.h:
* Source/WebKit/UIProcess/win/AutomationSessionClientWin.cpp: Copied from Source\WebKit\UIProcess\win\AutomationClientWin.cpp.
(WebKit::AutomationSessionClient::AutomationSessionClient):
(WebKit::AutomationSessionClient::close):
(WebKit::AutomationSessionClient::didReceiveAuthenticationChallenge):
(WebKit::AutomationSessionClient::requestNewPageWithOptions):
(WebKit::AutomationSessionClient::didDisconnectFromRemote):
(WebKit::AutomationSessionClient::retainWebView):
(WebKit::AutomationSessionClient::releaseWebView):
* Source/WebKit/UIProcess/win/AutomationSessionClientWin.h: Copied from Source\WebKit\UIProcess\win\AutomationClientWin.h.

Canonical link: https://commits.webkit.org/293149@main
